Parse data files and construct the HdrHistogram
However, it's still missing the important step of trying to deduplicate
timings from successive data files. At the 1 h collection interval I'm
using, most data from a data file will be redundant with that of the
previous data file.

Also note that the data comes in 15-min granularity, which is higher
than the claimed 5-min measurement interval. This suggests that these
are averages of several measurements.