Dedup finding the longest possible exact overlap
The previous algorithm tried to be too clever, and instead ended up
being too complicated, hard to reason about, quirky and, more
importantly, hard to get right. Just after the last fix, a new case
showed it to still be broken.

Replace it with finding the longest possible exact overlap (a slight
simplification from a maximal unique match, for general sequence
alignment). This is simple, predictable, and even performant (given the
short sequences we have to deal here).